Move libcontainerd/supervisor to daemon/internal/libcontainerd/supervisor
