+++ daemon/command/daemon.go
@@ -28,6 +28,7 @@ import (
 	"github.com/docker/docker/daemon/command/debug"
 	"github.com/docker/docker/daemon/command/trap"
 	"github.com/docker/docker/daemon/config"
+	"github.com/docker/docker/daemon/internal/libcontainerd/supervisor"
 	"github.com/docker/docker/daemon/listeners"
 	"github.com/docker/docker/daemon/pkg/plugin"
 	apiserver "github.com/docker/docker/daemon/server"
@@ -48,7 +49,6 @@ import (
 	"github.com/docker/docker/daemon/server/router/volume"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/otelutil"
-	"github.com/docker/docker/libcontainerd/supervisor"
 	dopts "github.com/docker/docker/opts"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/homedir"
