+++ daemon/internal/libcontainerd/supervisor/remote_daemon_windows.go
